庆阳项目/auto_test/requirements.txt
pytest>=7.4.0
requests>=2.31.0
allure-pytest>=2.13.2
openpyxl>=3.1.2
PyYAML>=6.0.1
